+++ database/migrations/2020_06_02_031736_create_employee_group_services_table.php
@@ -18,7 +18,7 @@ class CreateEmployeeGroupServicesTable extends Migration
             $table->bigIncrements('id');
 
             $table->unsignedInteger('employee_groups_id')->nullable();
-            $table->unsignedInteger('business_service_id')->nullable();
+            $table->unsignedBigInteger('business_service_id')->nullable();
 
             $table->foreign('employee_groups_id')->references('id')->on('employee_groups')->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('business_service_id')->references('id')->on('business_services')->onUpdate('cascade')->onDelete('cascade');
